+++ src/components/Animations/TumbleweedRollingAnimation.tsx
@@ -1,11 +1,19 @@
 import LottieView, { type LottieViewProps } from 'lottie-react-native';
-import React, { forwardRef, type ForwardRefRenderFunction } from 'react';
+import React, { forwardRef, useMemo, type ForwardRefRenderFunction } from 'react';
+import { useColorScheme } from 'react-native';
 import TumbleweedRolling from '@/assets/animations/tumbleweed-rolling.json';
+import { colouriseLottie } from '@/helpers/colors';
 
 type AnimationProps = Omit<LottieViewProps, 'source'>;
 
-const CouponsAnimation: ForwardRefRenderFunction<LottieView, AnimationProps> = (props, ref) => {
-  return <LottieView ref={ref} autoPlay loop {...props} source={TumbleweedRolling} />;
+const TumbleweedRollingAnimation: ForwardRefRenderFunction<LottieView, AnimationProps> = (
+  props,
+  ref,
+) => {
+  const colorScheme = useColorScheme();
+  const colorizedSource = useMemo(() => colouriseLottie(TumbleweedRolling, {}), [colorScheme]);
+
+  return <LottieView ref={ref} autoPlay loop {...props} source={colorizedSource} />;
 };
 
-export default forwardRef(CouponsAnimation);
+export default forwardRef(TumbleweedRollingAnimation);