+++ src/components/Onboarding/PaginationDot.tsx
@@ -1,4 +1,10 @@
-import Animated, { interpolate, interpolateColor, useAnimatedStyle } from 'react-native-reanimated';
+import { useColorScheme } from 'react-native';
+import Animated, {
+  interpolate,
+  interpolateColor,
+  SharedValue,
+  useAnimatedStyle,
+} from 'react-native-reanimated';
 import { theme } from '@/helpers/colors';
 
 const DOT_SIZE = 12;
@@ -9,7 +15,7 @@ const PaginationDot = ({
   containerWidth,
 }: {
   index: number;
-  animationValue: Animated.SharedValue<number>;
+  animationValue: SharedValue<number>;
   containerWidth: number;
 }) => {
   const inputRange = [
@@ -17,11 +23,15 @@ const PaginationDot = ({
     index * containerWidth,
     (index + 1) * containerWidth,
   ];
+  const colorScheme = useColorScheme();
   const animatedStyles = useAnimatedStyle(() => {
+    const isDark = colorScheme === 'dark';
     const colour = interpolateColor(
       animationValue.value,
       inputRange,
-      [theme.silverSand, '#C27803', theme.charlestonGreen], // TODO: fix crash when using tw.color
+      isDark
+        ? [theme.charlestonGreen, '#C27803', theme.silverSand]
+        : [theme.silverSand, '#C27803', theme.charlestonGreen], // TODO: fix crash when using tw.color
       'RGB',
     );
 
@@ -36,7 +46,7 @@ const PaginationDot = ({
       width: expand,
       backgroundColor: colour,
     };
-  });
+  }, [colorScheme]);
 
   return (
     <Animated.View