+++ src/context/i18n.tsx
@@ -1,13 +1,15 @@
 import dayjs from 'dayjs';
 import { createContext, useContext, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import LanguageBottomSheet from '@/components/Settings/LanguageBottomSheet';
 import { SYSTEM_LANGUAGE } from '@/i18n';
 import useSettingsStore, { SYSTEM_OPTION } from '@/stores/settings';
 
 const I18nContext = createContext<{
   language: string | null;
   ready: boolean;
-}>({ language: null, ready: false });
+  selectLanguage: () => void;
+}>({ language: null, ready: false, selectLanguage: () => { } });
 
 export const DEFAULT_LANGUAGE = process.env.EXPO_PUBLIC_DEFAULT_LANGUAGE || 'fr';
 
@@ -19,9 +21,11 @@ const useChosenLanguange = (language: string | null, setReady: (ready: boolean) 
   const { i18n } = useTranslation();
 
   useEffect(() => {
-    const chosenLanguage = !language || language === SYSTEM_OPTION ? SYSTEM_LANGUAGE : language;
-    const isLanguageSupported = Object.keys(i18n.options.resources || {}).includes(chosenLanguage);
-    const appliedLanguage = isLanguageSupported ? chosenLanguage : DEFAULT_LANGUAGE;
+    const selectedLanguage = !language || language === SYSTEM_OPTION ? SYSTEM_LANGUAGE : language;
+    const isLanguageSupported = Object.keys(i18n.options.resources || {}).includes(
+      selectedLanguage,
+    );
+    const appliedLanguage = isLanguageSupported ? selectedLanguage : DEFAULT_LANGUAGE;
     i18n.changeLanguage(appliedLanguage);
     dayjs.locale(appliedLanguage);
     setReady(true);
@@ -30,9 +34,16 @@ const useChosenLanguange = (language: string | null, setReady: (ready: boolean) 
 
 export const I18nProvider = ({ children }: { children: React.ReactNode }) => {
   const [ready, setReady] = useState<boolean>(false);
+  const [isSelecting, setSelecting] = useState<boolean>(false);
   const language = useSettingsStore((state) => state.language);
 
   useChosenLanguange(language, setReady);
 
-  return <I18nContext.Provider value={{ language, ready }}>{children}</I18nContext.Provider>;
+  return (
+    <I18nContext.Provider value={{ language, ready, selectLanguage: () => setSelecting(true) }}>
+      {children}
+
+      {isSelecting && <LanguageBottomSheet onClose={() => setSelecting(false)} />}
+    </I18nContext.Provider>
+  );
 };