+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Widescreen layout.
 - Newsletter link on socials bottom sheet.
 
 ### Fixed